IND-1502 make the historical node reject queries until it is ready
Copying the approach from `/druid/historical/v1/readiness` in
HistoricalResource.java.

The `ZkCoordinator` component is injected and `isStarted()` can be checked to
see if the node is ready for queries.

For the broker node, the ZkCoordinator component will be null, and not checked.
For the historical node, the ZkCoordinator component will be non-null, and
`isStarted()` will be checked during query initialization.

Addresses https://github.com/apache/incubator-druid/issues/6498